sync

+++ barretenberg/acir_tests/bootstrap.sh
@@ -12,8 +12,10 @@ function build {
     rm -rf acir_tests/{diamond_deps_0,workspace,workspace_default_member,regression_5045}
     # TODO(https://github.com/AztecProtocol/barretenberg/issues/1108): problem regardless the proof system used
     rm -rf acir_tests/regression_5045
-    # These honk tests just started failing...
-    rm -rf acir_tests/{verify_honk_proof,double_verify_honk_proof}
+    if [ "{CI25:-0}" = "1" ]; then
+      # These honk tests just started failing...
+      rm -rf acir_tests/{verify_honk_proof,double_verify_honk_proof}
+    fi
   fi
 
   # Compile 2 only compiles the tests.
